Make Hue::NormalizeMac public.
If the user wishes to connect to a specific bridge it makes sense for them to remeber the bridge by and scan for the bridge by comparing MAC addresses as these are more stable than IPs in an ipv4 DHCP network. In order for the MAC address comparison to be stable, the user should use the same normalisation scheme as the library. Exposing this function to the user makes their life easier.